glib: Show structure element standard attributes in demo
In poppler-glib-demo, display the values of the standard structure element
attributes. This also makes the grid view scrollable to make it possible to
see all the attributes when the window is small, and the horizontal GtkBox
is changed to a GtkPane, to make it possible to resize the tree in the left
column.